Update Parser Logic
