+++ tools/apiview/parsers/csharp-api-parser/CSharpAPIParser/CSharpAPIParser.csproj
@@ -11,6 +11,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="NuGet.Protocol" Version="6.10.1" />
     <PackageReference Include="System.CommandLine" Version="2.0.0-beta4.22272.1" />
   </ItemGroup>
