+++ tools/apiview/parsers/csharp-api-parser/CSharpAPIParser/Program.cs
@@ -4,6 +4,11 @@ using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Xml.Linq;
 using ApiView;
+using NuGet.Common;
+using NuGet.Packaging;
+using NuGet.Protocol;
+using NuGet.Protocol.Core.Types;
+using NuGet.Versioning;
 
 var inputOption = new Option<FileInfo>("--packageFilePath", "C# Package (.nupkg) file").ExistingOnly();
 inputOption.IsRequired = true;
@@ -21,13 +26,13 @@ var rootCommand = new RootCommand("Parse C# Package (.nupkg) to APIView Tokens")
     runAnalysis
 };
 
-rootCommand.SetHandler((FileInfo packageFilePath, DirectoryInfo outputDirectory, string outputFileName, bool runAnalysis) =>
+rootCommand.SetHandler(async (FileInfo packageFilePath, DirectoryInfo outputDirectory, string outputFileName, bool runAnalysis) =>
 {
     try
     {
         using (var stream = packageFilePath.OpenRead())
         {
-            HandlePackageFileParsing(stream, packageFilePath, outputDirectory, outputFileName, runAnalysis);
+            await HandlePackageFileParsing(stream, packageFilePath, outputDirectory, outputFileName, runAnalysis);
         }
     }
     catch (Exception ex)
@@ -39,12 +44,13 @@ rootCommand.SetHandler((FileInfo packageFilePath, DirectoryInfo outputDirectory,
 return rootCommand.InvokeAsync(args).Result;
 
 
-static void HandlePackageFileParsing(Stream stream, FileInfo packageFilePath, DirectoryInfo OutputDirectory, string outputFileName, bool runAnalysis)
+static async Task HandlePackageFileParsing(Stream stream, FileInfo packageFilePath, DirectoryInfo OutputDirectory, string outputFileName, bool runAnalysis)
 {
     ZipArchive? zipArchive = null;
     Stream? dllStream = stream;
     Stream? docStream = null;
     List<DependencyInfo>? dependencies = null;
+    string dependencyFilesTempDir = null;
 
     try
     {
@@ -94,12 +100,16 @@ static void HandlePackageFileParsing(Stream stream, FileInfo packageFilePath, Di
             }
         }
 
-        var assemblySymbol = CompilationFactory.GetCompilation(dllStream, docStream);
+        dependencyFilesTempDir = await ExtractNugetDependencies(dependencies).ConfigureAwait(false);
+        var dependencyFilePaths = Directory.EnumerateFiles(dependencyFilesTempDir, "*.dll", SearchOption.AllDirectories);
+        var assemblySymbol = CompilationFactory.GetCompilation(dllStream, docStream, dependencyFilePaths);
+
         if (assemblySymbol == null)
         {
             Console.Error.WriteLine($"PackageFile {packageFilePath.FullName} contains no Assembly Symbol.");
             return;
         }
+
         var parsedFileName = string.IsNullOrEmpty(outputFileName) ? assemblySymbol.Name : outputFileName;
         var treeTokenCodeFile = new CSharpAPIParser.TreeToken.CodeFileBuilder().Build(assemblySymbol, runAnalysis, dependencies);
         var gzipJsonTokenFilePath = Path.Combine(OutputDirectory.FullName, $"{parsedFileName}.json.tgz");
@@ -126,8 +136,11 @@ static void HandlePackageFileParsing(Stream stream, FileInfo packageFilePath, Di
     finally
     {
         zipArchive?.Dispose();
+        if (dependencyFilesTempDir != null && Directory.Exists(dependencyFilesTempDir))
+        {
+            Directory.Delete(dependencyFilesTempDir, true);
+        }
     }
-
 }
 
 static bool IsNuget(string name)
@@ -143,4 +156,49 @@ static bool IsNuspec(string name)
 static bool IsDll(string name)
 {
     return name.EndsWith(".dll", StringComparison.OrdinalIgnoreCase);
+}
+
+/// <summary>
+/// Resolves the NuGet package dependencies and extracts them to a temporary folder. It is the responsibility of teh caller to clean up the folder.
+/// </summary>
+/// <param name="dependencyInfos">The dependency infos</param>
+/// <returns>A temporary path where the dependency files were extracted.</returns>
+static async Task<string> ExtractNugetDependencies(List<DependencyInfo> dependencyInfos)
+{
+    string tempFolder = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString());
+    SourceCacheContext cache = new SourceCacheContext();
+    SourceRepository repository = Repository.Factory.GetCoreV3("https://api.nuget.org/v3/index.json");
+    try
+    {
+        FindPackageByIdResource resource = await repository.GetResourceAsync<FindPackageByIdResource>().ConfigureAwait(false);
+        foreach (var dep in dependencyInfos)
+        {
+            using (MemoryStream packageStream = new MemoryStream())
+            {
+                if (await resource.CopyNupkgToStreamAsync(
+                dep.Name,
+                new NuGetVersion(dep.Version),
+                packageStream,
+                cache,
+                NullLogger.Instance,
+                CancellationToken.None))
+                {
+                    using PackageArchiveReader reader = new PackageArchiveReader(packageStream);
+                    NuspecReader nuspec = reader.NuspecReader;
+                    var file = reader.GetFiles().FirstOrDefault(f => f.EndsWith(dep.Name + ".dll"));
+                    if (file != null)
+                    {
+                        var fileInfo = new FileInfo(file);
+                        var path = Path.Combine(tempFolder, dep.Name, fileInfo.Name);
+                        var tmp = reader.ExtractFile(file, path, NullLogger.Instance);
+                    }
+                }
+            }
+        }
+    }
+    finally
+    {
+        cache.Dispose();
+    }
+    return tempFolder;
 }
